+++ changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Add support for empty values in field definition
+
 ## v3.0
 ### Changed
 - Tiny but breaking changes (See upgrade hints in README.md)